Branch apps as mupdf.
Previously the example front end applications were called generically
'pdfview' and 'pdftool'. In preparation for release we'd like to brand
them more uniquely. Since the user-facing part of the tree is the mupdf
library, we intend to brand the releases mupdf.

In support of this, 'pdfview' becomes 'mupdf' and 'pdftool' becomes
'mupdftool'.